requirements.txt
python-dotenv==1.0.0
watchdog==4.0.0
psutil==5.9.6
